add html + part of degree (minutes)

+++ src/main/java/ru/nia/service/MultiplierService.java
@@ -3,23 +3,25 @@ package ru.nia.service;
 import org.apfloat.Apfloat;
 import org.apfloat.ApfloatMath;
 
+import java.math.BigDecimal;
+
 public class MultiplierService {
     private static final int precision = 80;
 
-    public Apfloat getCos(int val) {
+    public Apfloat getCos(BigDecimal val) {
         return ApfloatMath.cos(ApfloatMath.toRadians(new Apfloat(val, precision))).precision(precision);
     }
 
-    public Apfloat getSin(int val) {
+    public Apfloat getSin(BigDecimal val) {
         return ApfloatMath.sin(ApfloatMath.toRadians(new Apfloat(val, precision))).precision(precision);
     }
 
-    private Apfloat getK(int step) {
+    private Apfloat getK(BigDecimal step) {
         Apfloat apfloat = getCos(step);
         return apfloat.multiply(new Apfloat(2, precision));
     }
 
-    public Apfloat doRecur(Apfloat second, Apfloat first, int step) {
+    public Apfloat doRecur(Apfloat second, Apfloat first, BigDecimal step) {
         Apfloat res = getK(step);
 
         res = res.multiply(second).precision(precision);
